requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
decorator==5.2.1
executing==2.2.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
flask-paginate==2024.4.12
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Limiter
Flask-Mail==3.10.1
Flask-Talisman==1.1.0
python-dotenv==1.0.1
frozenlist==1.7.0
genai==2.1.0
google==3.0.0
google-auth==2.40.3
google-genai==1.31.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.6.4
openai==0.27.10
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-flask==1.3.0
regex==2025.7.34
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.20.1
